Custom widgets

Custom widgets

Besides the standard widgets, Interact provides a framework to define custom GUIs. This is currently possible with two approaches, the full featured @widget macro and the simple to use but more basic @manipulate macro.

The recipe macro

Widgets.@widget — Macro.

@widget(wdgname, func_call)

Special macro to create "recipes" for custom widgets. The @widget macro takes to argument, a variable name wdgname and a function call func_call. The function call is changed by the macro in several ways:

an extra line is added at the beginning to initiliaze a variable called wdgname::Widget that can be used to refer to the widget in the function body

all lines of the type sym::Symbol = expr are replaced with wdgname[sym] = @map(wdgname, expr), see Widgets.@map for more details

an extra line is added at the end to return wdgname

The macro then registers the function func_call and exports it. It also overloads the widget function with the following signature:

Widgets.widget(::Val{Symbol(func_name)}, args...; kwargs..) = func_name(args...; kwargs...)

source

Auxiliary functions

Widgets.@map — Macro.

@map(d, x)

Apply the expression x to the widget d, replacing e.g. symbol :s with the corresponding Observable observe(d[:s]). To use the value of some of d's children, use :s[]. Use $(:s) if you want the output to update automatically as soon as the value of observe(d[:s]) changes. In this context, _ refers to the whole widget. To use actual symbols, escape them with ^, as in ^(:a).

Examples

julia> using DataStructures, InteractBase, Observables

julia> t = Widgets.Widget{:test}(OrderedDict(:a => Observable(2), :b => slider(1:100), :c => button()));

This updates as soon as observe(t[:a]) or observe(t[:b]) change:

julia> Widgets.@map t $(:a) + $(:b)
Observables.Observable{Int64}("ob_31", 52, Any[])

whereas this only updates when button :c is pressed:

julia> Widgets.@map t ($(:c); :a[] + :b[])
Observables.Observable{Int64}("ob_33", 52, Any[])

source

Widgets.@output! — Macro.

@output!(d, x)

Computes Widgets.@map(d, x) and sets d.output to be the result (see Widgets.@map for more details). d.display is also set by default to match d.output. To have a custom display use @display!(d, expr) _after_ @output!(d, x)

source

Widgets.@display! — Macro.

@display!(d, x)

Computes Widgets.@map(d, x) and sets d.display to be the result (see Widgets.@map for more details).

source

Widgets.@layout — Macro.

@layout(x)

Returns a function, that takes as argument a widget d and replaces e.g. symbol :s with the corresponding subwidget d[:s]. In this context, _ refers to the whole widget. To use actual symbols, escape them with ^, as in ^(:a).

Examples

julia> using DataStructures, InteractBase, CSSUtil

julia> f = Widgets.@layout hbox(:b, CSSUtil.hskip(1em), :c);

julia> t = Widgets.Widget{:test}(OrderedDict(:b => slider(1:100), :c => button()));

julia> f(t);

source

Widgets.@layout! — Macro.

@layout!(d, x)

Set d.layout to match the result of Widgets.@layout(x). See Widgets.@layout for more information.

Examples

julia> using DataStructures, InteractBase, CSSUtil

julia> t = Widgets.Widget{:test}(OrderedDict(:b => slider(1:100), :c => button()));

julia> @layout! t hbox(:b, CSSUtil.hskip(1em), :c);

source

A simpler approach: the manipulate macro

InteractBase.@manipulate — Macro.

@manipulate expr

The @manipulate macro lets you play with any expression using widgets. expr needs to be a for loop. The for loop variable are converted to widgets using the widget function (ranges become slider, lists of options become togglebuttons, etc...). The for loop body is displayed beneath the widgets and automatically updated as soon as the widgets change value.

Use throttle = df to only update the output after a small time interval dt (useful if the update is costly as it prevents multiple updates when moving for example a slider).

Examples

using Colors

@manipulate for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end

@manipulate throttle = 0.1 for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end

@layout! can be used to adjust the layout of a manipulate block:

using Widgets, CSSUtil, WebIO

ui = @manipulate throttle = 0.1 for r = 0:.05:1, g = 0:.05:1, b = 0:.05:1
    HTML(string("<div style='color:#", hex(RGB(r,g,b)), "'>Color me</div>"))
end
@layout! ui dom"div"(_.display, vskip(2em), :r, :g, :b)
ui

source