Modifiers

Modifiers

Interact provides some modifiers to change an existing widget (or return a tranformed version of the widget):

InteractBase.tooltip! — Function.

tooltip!(wdg::AbstractWidget, tooltip; className = "")

Experimental. Add a tooltip to widget wdg. tooltip is the text that will be shown and className can be used to customize the tooltip, for example is-tooltip-bottom or is-tooltip-danger.

InteractBase.onchange — Function.

onchange(w::AbstractWidget, change = w[:changes])

Return a widget that's identical to w but only updates on change. For a slider it corresponds to releasing it and for a textbox it corresponds to losing focus.

Examples

sld = slider(1:100) |> onchange # update on release
txt = textbox("Write here") |> onchange # update on losing focuse